+++ packages/sdks/e2e/react-sdk-next-15-app/src/app/[[...slug]]/page.tsx
@@ -1,4 +1,5 @@
-import BuilderBlockWithClassName from '@/components/BuilderBlockWithClassName';
+import { builderBlockWithClassNameComponentConfig } from '@/components/BuilderBlockWithClassName';
+import ClientContent from '@/components/ClientContent';
 import {
   Content,
   _processContentResult,
@@ -6,48 +7,6 @@ import {
   getBuilderSearchParams,
 } from '@builder.io/sdk-react';
 import { getProps } from '@sdk/tests';
-
-const builderBlockWithClassNameCustomComponent = {
-  name: 'BuilderBlockWithClassName',
-  component: BuilderBlockWithClassName,
-  isRSC: true,
-  shouldReceiveBuilderProps: {
-    builderBlock: true,
-    builderContext: true,
-    builderComponents: true,
-  },
-  inputs: [
-    {
-      name: 'content',
-      type: 'uiBlocks',
-      defaultValue: [
-        {
-          '@type': '@builder.io/sdk:Element',
-          '@version': 2,
-          id: 'builder-c6e179528dee4e62b337cf3f85d6496f',
-          component: {
-            name: 'Text',
-            options: {
-              text: 'Enter some text...',
-            },
-          },
-          responsiveStyles: {
-            large: {
-              display: 'flex',
-              flexDirection: 'column',
-              position: 'relative',
-              flexShrink: '0',
-              boxSizing: 'border-box',
-              marginTop: '20px',
-              lineHeight: 'normal',
-              height: 'auto',
-            },
-          },
-        },
-      ],
-    },
-  ],
-};
 
 interface PageProps {
   params: Promise<{
@@ -76,7 +35,11 @@ export default async function Page(props: PageProps) {
     );
   }
 
-  builderProps.customComponents = [builderBlockWithClassNameCustomComponent];
+  builderProps.customComponents = [builderBlockWithClassNameComponentConfig];
+
+  if (urlPath === '/variant-containers') {
+    return <ClientContent {...builderProps} />;
+  }
 
   return <Content {...builderProps} />;
 }