+++ packages/sdks/src/scripts/init-editing.ts
@@ -54,6 +54,7 @@ export const setupBrowserForEditing = (
           supportsCustomBreakpoints: true,
           modelName:options.modelName,
           apiKey:options.apiKey,
+          blockLevelPersonalization: true,
         },
       },
       '*'