+++ packages/webcomponents/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.0.1
+- update `@builder.io/sdk` to `^6.0.4` and `@builder.io/react` to `^8.0.7`
+- add check around builder block id so that widgets and blocks do not crash.
+
 ### 2.0.0
 - update `@builder.io/sdk` to `^5.0.0` and `@builder.io/react` to `^7.0.0`
 - send SDK name and version information to Content Editor when editing content.